requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
dash>=2.0.0
pyyaml>=5.1
